# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Watch Video & Earn API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthCredentials {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface AuthResponse {
  userId: string;
  email: string;
  points: number;
  token: string;
}

export interface UserProfile {
  userId: string;
  email: string;
  points: number;
  /** @nullable */
  lastDailyBonus: string | null;
  canClaimDailyBonus: boolean;
}

export interface CashoutResponse {
  message: string;
  pointsRedeemed: number;
  remainingPoints: number;
}

export interface ErrorResponse {
  error: string;
}