# lib/api-zod/src/generated/types/authResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Watch Video & Earn API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthResponse {
  userId: string;
  email: string;
  points: number;
  token: string;
}