# lib/api-zod/src/generated/types/cashoutResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Watch Video & Earn API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CashoutResponse {
  message: string;
  pointsRedeemed: number;
  remainingPoints: number;
}