# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Watch Video & Earn API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  userId: string;
  email: string;
  points: number;
  /** @nullable */
  lastDailyBonus: string | null;
  canClaimDailyBonus: boolean;
}